requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
numpy==1.24.3
scikit-learn==1.3.2
pandas==2.1.4
autopep8==2.0.4
black==23.12.1
pylint==3.0.3
radon==6.0.1
astroid==3.0.2
jinja2==3.1.2
werkzeug==3.0.1
gunicorn==21.2.0
